+++ src/main/java/com/library/services/MemberService.java
@@ -1,11 +1,14 @@
 package com.library.services;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import com.library.dto.MemberRequest;
+import com.library.exceptions.MemberNotFoundException;
 import com.library.models.entities.Member;
 import com.library.models.repositories.MemberRepo;
 
@@ -17,16 +20,38 @@ public class MemberService {
     @Autowired
     private MemberRepo memberRepo;
 
-    public Member saveMember(Member member) {
+    public Member createMember(MemberRequest memberRequest) {
+        Member member = Member.builder().memberName(memberRequest.getMemberName())
+                .dateCreated(memberRequest.getDateCreated()).dateUpdated(memberRequest.getDateUpdated()).build();
+
         return memberRepo.save(member);
     }
 
-    public Member updateMember(Member member) {
-        return memberRepo.save(member);
+    public Member updateMember(MemberRequest memberRequest) throws MemberNotFoundException {
+        Optional<Member> member = memberRepo.findById(memberRequest.getId());
+
+        if (!member.isPresent()) {
+            throw new MemberNotFoundException("Member with id " + member + " not found");
+        } else {
+            member.get().setMemberName(memberRequest.getMemberName());
+            member.get().setDateUpdated(memberRequest.getDateUpdated());
+
+            return memberRepo.save(member.get());
+        }
     }
 
-    public Iterable<Member> findAllMember() {
-        return memberRepo.findAll();
+    // Update Just Specific Field
+    public Member updateSpecificField() {
+        return null;
+    }
+
+    public Iterable<Member> findAllMember() throws MemberNotFoundException {
+        Iterable<Member> members = memberRepo.findAll();
+        if (members == null) {
+            throw new MemberNotFoundException("No one member in here !! ");
+        } else {
+            return members;
+        }
     }
 
     public Member findOneMember(Long id) {
@@ -38,8 +63,15 @@ public class MemberService {
         return member.get();
     }
 
-    public void deleteMemberById(Long id) {
-        memberRepo.deleteById(id);
+    public void deleteMemberById(Long id) throws MemberNotFoundException {
+        Optional<Member> member = memberRepo.findById(id);
+
+        if (!member.isPresent()) {
+            throw new MemberNotFoundException("Member with id " + member + " not found");
+        } else {
+            memberRepo.deleteById(id);
+        }
+
     }
 
     public Iterable<Member> findByMemberName(String memberName, Pageable pageable) {
